backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
asyncpg
psycopg2-binary
alembic
requests
beautifulsoup4
lxml
google-generativeai
